Renamed files, added Flight Trajectory class
