+++ src/RocketMath.h
@@ -1,7 +1,7 @@
 #ifndef ROCKET_MATH_H
 #define ROCKET_MATH_H
 
-#include "map/position.h"
+#include "map/Position.h"
 
 namespace RocketMath {
     double haversine2D(const Position& p1, const Position& p2);
